Main window: Move create_application to utils
